Issue #6: Candidates detection - Gaussian pyramids (thresholds + grouping) -> Changed for linear structuring elements analysis
Function improved for more accurate detection of candidates. Main changes:
* Define d1 as 0.6*d2 instead of 0.7*d2 to give more flexibility to the detection range
* Number of scales increased so that more sizes are covered
All ground truth images were analyzed visually to check that every mass is in at least one of the scales. For most of the images the detection is satisfactory. In some cases (4 cases) the detected candidate is slightly smaller than the real mass.